updated CLI

+++ src/main/java/com/techelevator/dao/ReservationDAO.java
@@ -9,7 +9,7 @@ import com.techelevator.model.Reservation;
 
 public interface ReservationDAO
 {
-	int createReservation(int siteId, String name, LocalDate fromDate, LocalDate toDate) throws IllegalArgumentException, InvalidKeyException;
+	int createReservation(long siteId, String name, LocalDate fromDate, LocalDate toDate) throws IllegalArgumentException, InvalidKeyException;
 	List<Reservation> getReservationsByStartDateForCampground(Campground campground);
 
 }
